Move shared code to the shared module
