+++ src/metaswitch/clearwater/cluster_manager/test/test_node_failure.py
@@ -34,7 +34,7 @@
 
 
 from mock import patch
-from .mock_python_etcd import EtcdFactory
+from metaswitch.clearwater.etcd_shared.test.mock_python_etcd import EtcdFactory
 from metaswitch.clearwater.cluster_manager.etcd_synchronizer \
     import EtcdSynchronizer
 from metaswitch.clearwater.cluster_manager.null_plugin import NullPlugin
