+++ src/metaswitch/clearwater/config_manager/test/test_base.py
@@ -34,7 +34,7 @@
 
 
 import unittest
-from metaswitch.clearwater.cluster_manager.test.mock_python_etcd import EtcdFactory
+from metaswitch.clearwater.etcd_shared.test.mock_python_etcd import EtcdFactory
 from metaswitch.clearwater.config_manager.etcd_synchronizer import \
     EtcdSynchronizer
 from .plugin import TestPlugin
